+++ lib/axn/webhooks.rb
@@ -7,6 +7,7 @@ require_relative "webhooks/version"
 require_relative "webhooks/request"
 require_relative "webhooks/signature"
 require_relative "webhooks/resolvers"
+require_relative "webhooks/verify"
 
 module Axn
   module Webhooks